for self testing

+++ integtest/test.sh
@@ -1,17 +1,23 @@
 #!/usr/bin/env bash
 set -ue
 
+BASE_DIR=$(cd `dirname $0`; pwd)
+echo ${BASE_DIR}
+echo "usage: <bin> [-d | --debug]"
+
 clear_last_diff_files() {
-    if [ -z "ls *.spark" ]; then
-        rm *.spark
-    fi
-    if [ -z "ls *.tidb" ]; then
-        rm *.tidb
-    fi
+    for f in ./*.spark; do
+        [ -e "$f" ] && rm *.spark
+        break
+    done
+    for f in ./*.tidb; do
+        [ -e "$f" ] && rm *.tidb
+        break
+    done
 }
 
 check_tpch_dir_is_present() {
-    if [ ! -d "tpch" ]; then
+    if [ ! -d "${BASE_DIR}/tpch" ]; then
         echo "tpch is not present. You have to clone it to your local machine."
         echo "this script is required to generate and load tpch database to TiDB cluster."
         echo "git clone https://github.com/zhexuany/tispark_tpch tpch"
@@ -35,19 +41,17 @@ clear_last_diff_files
 check_tpch_dir_is_present
 check_tpch_data_is_loaded
 
-BASEDIR=$(dirname "$0")
 CLASS="com.pingcap.spark.TestFramework"
 
-cp ${BASEDIR}/conf/tispark_config.properties "${SPARK_HOME}"/conf/tispark_config.properties
-spark_debug_opt="-agentlib:jdwp=transport=dt_socket,server=y,address=5005,suspend=y -Dtest.mode=Test"
-spark_test_opt="-Dtest.mode=Test"
+cp ${BASE_DIR}/conf/tispark_config.properties.template ${SPARK_HOME}/conf/tispark_config.properties
+spark_debug_opt="-Xdebug -Xrunjdwp:server=y,transport=dt_socket,address=5005,suspend=y"
+spark_test_opt=""
 
-spark_cmd="${SPARK_HOME}/bin/spark-submit --class $CLASS ${BASEDIR}/lib/* --driver-java-options"
+spark_cmd="${SPARK_HOME}/bin/spark-submit --class ${CLASS} ${BASE_DIR}/lib/* --driver-java-options"
 if [[ "$@" = *--debug ]] || [[ "$@" = *-d ]]; then
     echo "debuging..."
-    $spark_cmd $spark_debug_opt | grep "*.sql result"
+    $spark_cmd $spark_debug_opt
 else
     echo "testing...."
-    $spark_cmd $spark_test_opt 2>&1 | grep "result:"
+    $spark_cmd $spark_test_opt 2>&1 | grep "result:\|time:"
 fi
-
